modify_rxor
